move around some modules
